+++ src/caching/github-actions-cache.ts
@@ -6,73 +6,24 @@
 
 import * as pathlib from 'path';
 import * as fs from 'fs/promises';
+import * as https from 'https';
 import {createHash} from 'crypto';
-import * as cacheUtils from '@actions/cache/lib/internal/cacheUtils.js';
-import {createTar, extractTar} from '@actions/cache/lib/internal/tar.js';
-import {retryTypedResponse} from '@actions/cache/lib/internal/requestUtils.js';
-import {
-  saveCache,
-  downloadCache,
-} from '@actions/cache/lib/internal/cacheHttpClient.js';
-import {HttpClient} from '@actions/http-client';
-import {BearerCredentialHandler} from '@actions/http-client/auth.js';
-import {isSuccessStatusCode} from '@actions/cache/lib/internal/requestUtils.js';
 import {scriptReferenceToString} from '../script.js';
 import {getScriptDataDir} from '../util/script-data-dir.js';
+import {execFile} from 'child_process';
+import {createReadStream, createWriteStream} from 'fs';
 
-import type {CompressionMethod} from '@actions/cache/lib/internal/constants.js';
-import type {
-  ReserveCacheRequest,
-  ReserveCacheResponse,
-  ITypedResponseWithError,
-  ArtifactCacheEntry,
-} from '@actions/cache/lib/internal/contracts.js';
+import type * as http from 'http';
 import type {Cache, CacheHit} from './cache.js';
 import type {ScriptReference, ScriptStateString} from '../script.js';
 import type {Logger} from '../logging/logger.js';
 import type {RelativeEntry} from '../util/glob.js';
-import {Result} from '../error.js';
-
-// TODO(aomarks) Consider dropping the dependency on @actions/cache by writing
-// our own implementation. See https://github.com/google/wireit/issues/107 for
-// details.
-//
-// The public API provided by the @actions/cache package doesn't exactly meet
-// our needs, because it automatically uses the file paths that are included in
-// the tarball as part of the cache entry version (see
-// https://github.com/actions/toolkit/blob/7654d97eb6c4a3d564f036a2d4a783ae9105ec07/packages/cache/src/internal/cacheHttpClient.ts#L70),
-// and implements globbing differently.
-//
-// We want complete control over our cache key, instead of having it be
-// generated automatically based on file paths.
-//
-// For this reason, we are reaching into the "internal/" directory of this
-// package to get more control. This is bad because those modules could change
-// at any time, which is why we currently have a strict ("=") version pin in our
-// package.json.
-//
-// The @actions/cache package is also our largest dependency by far. It's 22MB,
-// and adds 63 transitive dependencies.
-//
-// The logic in here could be re-implemented in a fairly minimal way. The main
-// tricky part is the way it handles tarball generation across platforms
-// (https://github.com/actions/toolkit/blob/7654d97eb6c4a3d564f036a2d4a783ae9105ec07/packages/cache/src/internal/tar.ts).
+import type {Result} from '../error.js';
 
 /**
  * Caches script output to the GitHub Actions caching service.
  */
 export class GitHubActionsCache implements Cache {
-  static #compressionMethodPromise?: Promise<CompressionMethod>;
-
-  static get #compressionMethod(): Promise<CompressionMethod> {
-    // We only need to do this once per Wireit process. It's expensive because
-    // it spawns a child process to test the tar version.
-    if (this.#compressionMethodPromise === undefined) {
-      this.#compressionMethodPromise = cacheUtils.getCompressionMethod();
-    }
-    return this.#compressionMethodPromise;
-  }
-
   readonly #baseUrl: string;
   readonly #authToken: string;
   readonly #logger: Logger;
@@ -101,6 +52,8 @@ export class GitHubActionsCache implements Cache {
     // they are _not_ provided to regular "run" scripts. For this reason, we
     // re-export those variables so that all "run" scripts can access them using
     // the "google/wireit@setup-github-actions-caching/v1" re-usable workflow.
+    //
+    // https://github.com/actions/toolkit/blob/500d0b42fee2552ae9eeb5933091fe2fbf14e72d/packages/cache/src/internal/cacheHttpClient.ts#L38
     const baseUrl = process.env['ACTIONS_CACHE_URL'];
     if (!baseUrl) {
       return {
@@ -129,6 +82,7 @@ export class GitHubActionsCache implements Cache {
       };
     }
 
+    // https://github.com/actions/toolkit/blob/500d0b42fee2552ae9eeb5933091fe2fbf14e72d/packages/cache/src/internal/cacheHttpClient.ts#L63
     const authToken = process.env['ACTIONS_RUNTIME_TOKEN'];
     if (!authToken) {
       return {
@@ -157,21 +111,35 @@ export class GitHubActionsCache implements Cache {
       return undefined;
     }
 
-    const compressionMethod = await GitHubActionsCache.#compressionMethod;
-    const version = this.#computeVersion(stateStr, compressionMethod);
-    const location = await this.#checkForCacheEntry(
-      script,
-      this.#computeCacheKey(script),
-      version
-    );
-    if (location === undefined) {
-      // No cache hit.
+    const version = this.#computeVersion(stateStr);
+    const key = this.#computeCacheKey(script);
+    const url = new URL('_apis/artifactcache/cache', this.#baseUrl);
+    url.searchParams.set('keys', key);
+    url.searchParams.set('version', version);
+
+    const {req, resPromise} = this.#request(url);
+    req.end();
+    const res = await resPromise;
+
+    if (res.statusCode === /* No Content */ 204) {
       return undefined;
     }
-    return new GitHubActionsCacheHit(
-      location,
-      compressionMethod,
-      this.#emptyDirectoriesManifestPath(script, version)
+
+    if (isOk(res)) {
+      const {archiveLocation} = JSON.parse(await readBody(res)) as {
+        archiveLocation: string;
+      };
+      return new GitHubActionsCacheHit(script, archiveLocation);
+    }
+
+    if (res.statusCode === /* Too Many Requests */ 429) {
+      this.#onRateLimit(script);
+      return;
+    }
+
+    throw new Error(
+      `GitHub Cache check HTTP ${String(res.statusCode)} error: ` +
+        (await readBody(res))
     );
   }
 
@@ -184,121 +152,193 @@ export class GitHubActionsCache implements Cache {
       return false;
     }
 
-    // We're going to build a tarball. We do this by passing paths to the "tar"
-    // command. When we pass a directory to "tar", all of its children are
-    // implicitly included. This is a problem, because sometimes we want to add
-    // a directory, but only a subset of its children (or none of them), because
-    // of exclusion patterns in the "output" globs.
-    //
-    // To work around this problem, we never pass directories to "tar". Instead,
-    // we enumerate every specific file. This works fine, except for empty
-    // directories. If a directory was explicitly listed for inclusion, and it
-    // happens to be empty, we should cache an empty directory.
-    //
-    // So we need special handling for empty directories. If there are empty
-    // directories, we create a special "empty directories manifest" file, and
-    // include that in the tarball at a predictable location (inside the
-    // ".wireit" folder so that it won't collide with a user file). Then when we
-    // untar, we check for that manifest, create any missing directories listed
-    // inside it, and delete the manifest.
-
-    const files = new Set<string>();
-    const emptyDirs = new Set<string>();
-    {
-      const nonEmptyDirs = new Set<string>();
-      for (const {path, dirent} of relativeFiles) {
-        const absPath = pathlib.join(script.packageDir, path);
-        if (dirent.isDirectory()) {
-          // Initially assume every directory might be empty. We might see a
-          // directory entry after a file that's in it. We'll filter it down
-          // after this loop.
-          emptyDirs.add(absPath);
-        } else {
-          files.add(absPath);
-          // Add all parent directories of this file to set of non-empty
-          // directories.
-          let cur = pathlib.dirname(absPath);
-          while (!nonEmptyDirs.has(cur)) {
-            // Note if we've already added the child, we must have already added
-            // all of its parents too.
-            nonEmptyDirs.add(cur);
-            cur = pathlib.dirname(cur);
-          }
-        }
-      }
-      for (const nonEmptyDir of nonEmptyDirs) {
-        emptyDirs.delete(nonEmptyDir);
-      }
-    }
-
-    const compressionMethod = await GitHubActionsCache.#compressionMethod;
-    const version = this.#computeVersion(stateStr, compressionMethod);
-
-    let emptyDirsManifestPath: string | undefined;
-    if (emptyDirs.size > 0) {
-      const emptyDirsManifest = JSON.stringify([...emptyDirs]);
-      emptyDirsManifestPath = this.#emptyDirectoriesManifestPath(
-        script,
-        version
-      );
-      await fs.mkdir(pathlib.dirname(emptyDirsManifestPath), {recursive: true});
-      await fs.writeFile(emptyDirsManifestPath, emptyDirsManifest, {
-        encoding: 'utf8',
-      });
-      files.add(emptyDirsManifestPath);
-    }
-
-    const tarballPath = await this.#makeTarball([...files], compressionMethod);
+    const absFiles = relativeFiles.map((rel) =>
+      pathlib.join(script.packageDir, rel.path)
+    );
+    const tempDir = await makeTempDir(script);
     try {
-      const tarBytes = cacheUtils.getArchiveFileSizeInBytes(tarballPath);
-      // Reference: https://github.com/actions/toolkit/blob/f8a69bc473af4a204d0c03de61d5c9d1300dfb17/packages/cache/src/cache.ts#L174
-      const GB = 1024 * 1024 * 1024;
-      const maxBytes = 10 * GB;
-      if (tarBytes > maxBytes) {
-        this.#logger.log({
-          script,
-          type: 'info',
-          detail: 'generic',
-          message:
-            `Output was too big to be cached: ` +
-            `${Math.round(tarBytes / GB)}GB > ` +
-            `${Math.round(maxBytes / GB)}GB.`,
-        });
-        return false;
-      }
-      const id = await this.#reserveCacheEntry(
+      const tarballPath = await this.#makeTarball([...absFiles], tempDir);
+      return await this.#reserveUploadAndCommitTarball(
         script,
-        this.#makeAuthenticatedHttpClient(),
-        this.#computeCacheKey(script),
-        version,
-        tarBytes
+        stateStr,
+        tarballPath
       );
-      // It's likely that we'll occasionally fail to reserve an entry and get
-      // undefined here, especially when running multiple GitHub Action jobs in
-      // parallel with the same scripts, because there is a window of time
-      // between calling "get" and "set" on the cache in which another worker
-      // could have reserved the entry before us. Non fatal, just don't save.
-      if (id !== undefined) {
-        try {
-          await saveCache(id, tarballPath);
-        } catch (error) {
-          if (/\W429\W/.test((error as Error).message)) {
-            this.#onRateLimit(script);
-            return false;
-          }
-          throw error;
-        }
-      }
     } finally {
-      // Delete the tarball.
-      const tarballDeleted = cacheUtils.unlinkFile(tarballPath);
-      // Also delete the empty directories manifest file.
-      if (emptyDirsManifestPath !== undefined) {
-        await fs.unlink(emptyDirsManifestPath);
-      }
-      await tarballDeleted;
+      await fs.rm(tempDir, {recursive: true});
+    }
+  }
+
+  /**
+   * @returns True if we reserved, uploaded, and committed the tarball. False if
+   * we gave up due to a rate limit error.
+   * @throws If an unexpected HTTP error occured.
+   */
+  async #reserveUploadAndCommitTarball(
+    script: ScriptReference,
+    stateStr: ScriptStateString,
+    tarballPath: string
+  ): Promise<boolean> {
+    const tarballStats = await fs.stat(tarballPath);
+    const tarballBytes = tarballStats.size;
+    // Reference:
+    // https://github.com/actions/toolkit/blob/f8a69bc473af4a204d0c03de61d5c9d1300dfb17/packages/cache/src/cache.ts#L174
+    const GB = 1024 * 1024 * 1024;
+    const maxBytes = 10 * GB;
+    if (tarballBytes > maxBytes) {
+      this.#logger.log({
+        script,
+        type: 'info',
+        detail: 'generic',
+        message:
+          `Output was too big to be cached: ` +
+          `${Math.round(tarballBytes / GB)}GB > ` +
+          `${Math.round(maxBytes / GB)}GB.`,
+      });
+      return false;
+    }
+    const id = await this.#reserveCacheEntry(
+      script,
+      this.#computeCacheKey(script),
+      this.#computeVersion(stateStr),
+      tarballBytes
+    );
+    // It's likely that we'll occasionally fail to reserve an entry and get
+    // undefined here, especially when running multiple GitHub Action jobs in
+    // parallel with the same scripts, because there is a window of time between
+    // calling "get" and "set" on the cache in which another worker could have
+    // reserved the entry before us. Non fatal, just don't save.
+    if (id === undefined) {
+      return false;
+    }
+    if (!(await this.#upload(script, id, tarballPath, tarballBytes))) {
+      return false;
+    }
+    if (!(await this.#commit(script, id, tarballBytes))) {
+      return false;
     }
     return true;
+  }
+
+  /**
+   * @returns True if we uploaded, false if we gave up due to a rate limit error.
+   * @throws If an unexpected HTTP error occured.
+   */
+  async #upload(
+    script: ScriptReference,
+    id: number,
+    tarballPath: string,
+    tarballBytes: number
+  ): Promise<boolean> {
+    const url = new URL(`_apis/artifactcache/caches/${id}`, this.#baseUrl);
+    // Reference:
+    // https://github.com/actions/toolkit/blob/500d0b42fee2552ae9eeb5933091fe2fbf14e72d/packages/cache/src/options.ts#L59
+    const maxChunkSize = 32 * 1024 * 1024;
+    const tarballHandle = await fs.open(tarballPath, 'r');
+    let offset = 0;
+    try {
+      // TODO(aomarks) Chunks could be uploaded in parallel.
+      while (offset < tarballBytes) {
+        const chunkSize = Math.min(tarballBytes - offset, maxChunkSize);
+        const start = offset;
+        const end = offset + chunkSize - 1;
+        offset += maxChunkSize;
+
+        const tarballChunkStream = createReadStream(tarballPath, {
+          fd: tarballHandle.fd,
+          start,
+          end,
+          autoClose: false,
+        });
+
+        const opts = {
+          method: 'PATCH',
+          headers: {
+            'content-type': 'application/octet-stream',
+            'content-range': `bytes ${start}-${end}/*`,
+          },
+        };
+        const {req, resPromise} = this.#request(url, opts);
+        tarballChunkStream.pipe(req);
+        tarballChunkStream.on('close', () => {
+          req.end();
+        });
+
+        const res = await resPromise;
+
+        if (res.statusCode === /* Too Many Requests */ 429) {
+          this.#onRateLimit(script);
+          return false;
+        }
+
+        if (!isOk(res)) {
+          throw new Error(
+            `GitHub Cache upload HTTP ${String(
+              res.statusCode
+            )} error: ${await readBody(res)}\nopts: ${JSON.stringify(opts)}`
+          );
+        }
+      }
+      return true;
+    } finally {
+      await tarballHandle.close();
+    }
+  }
+
+  /**
+   * @returns True if we committed, false if we gave up due to a rate limit error.
+   * @throws If an unexpected HTTP error occured.
+   */
+  async #commit(
+    script: ScriptReference,
+    id: number,
+    tarballBytes: number
+  ): Promise<boolean> {
+    const url = new URL(
+      `_apis/artifactcache/caches/${String(id)}`,
+      this.#baseUrl
+    );
+    const reqBody = JSON.stringify({
+      size: tarballBytes,
+    });
+    const {req, resPromise} = this.#request(url, {
+      method: 'POST',
+      headers: {
+        'content-type': 'application/json',
+      },
+    });
+    req.end(reqBody);
+    const res = await resPromise;
+
+    if (res.statusCode === /* Too Many Requests */ 429) {
+      this.#onRateLimit(script);
+      return false;
+    }
+
+    if (!isOk(res)) {
+      throw new Error(
+        `GitHub Cache commit HTTP ${String(
+          res.statusCode
+        )} error: ${await readBody(res)}`
+      );
+    }
+
+    return true;
+  }
+
+  #request(
+    url: URL,
+    options?: http.RequestOptions
+  ): {req: http.ClientRequest; resPromise: Promise<http.IncomingMessage>} {
+    return request(url, {
+      ...options,
+      headers: {
+        // https://github.com/actions/toolkit/blob/500d0b42fee2552ae9eeb5933091fe2fbf14e72d/packages/cache/src/internal/cacheHttpClient.ts#L55
+        accept: 'application/json;api-version=6.0-preview.1',
+        // https://github.com/actions/toolkit/blob/500d0b42fee2552ae9eeb5933091fe2fbf14e72d/packages/http-client/src/auth.ts#L46
+        authorization: `Bearer ${this.#authToken}`,
+        ...options?.headers,
+      },
+    });
   }
 
   /**
@@ -318,31 +358,18 @@ export class GitHubActionsCache implements Cache {
     this.#hitRateLimit = true;
   }
 
-  #emptyDirectoriesManifestPath(
-    script: ScriptReference,
-    version: string
-  ): string {
-    return pathlib.join(
-      getScriptDataDir(script),
-      `github-cache-empty-directories-manifest-${version}.json`
-    );
-  }
-
   #computeCacheKey(script: ScriptReference): string {
     return createHash('sha256')
       .update(scriptReferenceToString(script))
       .digest('hex');
   }
 
-  #computeVersion(
-    stateStr: ScriptStateString,
-    compressionMethod: CompressionMethod
-  ): string {
+  #computeVersion(stateStr: ScriptStateString): string {
     return createHash('sha256')
       .update(
         [
           stateStr,
-          compressionMethod, // e.g. zstd, gzip
+          'gzip', // e.g. zstd, gzip
           // The ImageOS environment variable tells us which operating system
           // version is being used for the worker VM (e.g. "ubuntu20",
           // "macos11"). We already include process.platform in ScriptState, but
@@ -352,24 +379,13 @@ export class GitHubActionsCache implements Cache {
           // could consider including, but it probably changes frequently and is
           // unlikely to affect output, so we prefer the higher cache hit rate.
           process.env.ImageOS ?? '',
+          // Versioning salt:
+          //   - <omitted>: Initial version.
+          //   - 2: Removed empty directories manifest.
+          '2',
         ].join('\x1E') // ASCII record seperator
       )
       .digest('hex');
-  }
-
-  /**
-   * Make an HTTP client which always sends "Authorization: Bearer <token>"
-   * header.
-   */
-  #makeAuthenticatedHttpClient(): HttpClient {
-    const bearerCredentialHandler = new BearerCredentialHandler(
-      this.#authToken
-    );
-    return new HttpClient('actions/cache', [bearerCredentialHandler], {
-      headers: {
-        Accept: 'application/json;api-version=6.0-preview.1',
-      },
-    });
   }
 
   /**
@@ -377,126 +393,112 @@ export class GitHubActionsCache implements Cache {
    *
    * @returns The full path to the tarball file on disk.
    */
-  async #makeTarball(
-    paths: string[],
-    compressionMethod: CompressionMethod
-  ): Promise<string> {
-    const folder = await cacheUtils.createTempDirectory();
-    await createTar(folder, paths, compressionMethod);
-    const path = pathlib.join(
-      folder,
-      cacheUtils.getCacheFileName(compressionMethod)
-    );
-    return path;
-  }
-
-  /**
-   * Check for a cache entry.
-   *
-   * @returns A tarball URL if this cache entry exists, or undefined if it does
-   * not exist.
-   */
-  async #checkForCacheEntry(
-    script: ScriptReference,
-    key: string,
-    version: string
-  ): Promise<string | undefined> {
-    const httpClient = this.#makeAuthenticatedHttpClient();
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const response: ITypedResponseWithError<ArtifactCacheEntry> =
-      await retryTypedResponse('getCacheEntry', async () =>
-        httpClient.getJson<ArtifactCacheEntry>(
-          /** For docs on this API, see {@link FakeGitHubActionsCacheServer} */
-          `${this.#baseUrl}_apis/artifactcache/cache?keys=${encodeURIComponent(
-            key
-          )}&version=${version}`
-        )
+  async #makeTarball(paths: string[], tempDir: string): Promise<string> {
+    // Create a manifest file so that we can pass a large number of files to
+    // tar.
+    const manifestPath = pathlib.join(tempDir, 'manifest.txt');
+    await fs.writeFile(manifestPath, paths.join('\n'), 'utf8');
+    const tarballPath = pathlib.join(tempDir, 'cache.tgz');
+    await new Promise<void>((resolve, reject) => {
+      execFile(
+        'tar',
+        [
+          // Use the newer standardized tar format.
+          '--posix',
+          // Use gzip compression.
+          //
+          // TODO(aomarks) zstd is faster and has better performance, but it's
+          // availability is unreliable, and appears to have a bug on Windows
+          // (https://github.com/actions/cache/issues/301). Investigate and
+          // enable if easy.
+          '--gzip',
+          '--create',
+          '--file',
+          tarballPath,
+          // Use absolute paths (note we use the short form because the long
+          // form is --absolute-names on GNU tar, but --absolute-paths on BSD
+          // tar).
+          '-P',
+          // We have a complete list of files and directories, so we don't need
+          // or want tar to automatically expand directories. This also allows
+          // us to create empty directories, even if they aren't actually empty
+          // on disk.
+          '--no-recursion',
+          '--files-from',
+          manifestPath,
+        ],
+        (error: unknown) => {
+          if (error != null) {
+            reject(`tar error: ${String(error)}`);
+          } else {
+            resolve();
+          }
+        }
       );
-    if (response.statusCode === /* No Content */ 204) {
-      return undefined;
-    }
-    if (response.statusCode === /* Too Many Requests */ 429) {
-      this.#onRateLimit(script);
-      return undefined;
-    }
-    if (
-      !isSuccessStatusCode(response.statusCode) ||
-      response.error !== undefined ||
-      response.result?.archiveLocation === undefined
-    ) {
-      throw new Error(
-        `Error getting cache entry: ${response.statusCode} ${
-          response.error?.message ?? '<no error message>'
-        }`
-      );
-    }
-    return response.result.archiveLocation;
+    });
+    return tarballPath;
   }
 
   /**
    * Reserve a cache entry.
    *
    * @returns A numeric cache id the cache entry was reserved for us, or
-   * undefined if the cache entry was already reserved.
+   * undefined if the cache entry was already reserved, or a rate limit error
+   * occured.
    */
   async #reserveCacheEntry(
     script: ScriptReference,
-    httpClient: HttpClient,
     key: string,
     version: string,
     cacheSize: number
   ): Promise<number | undefined> {
-    const request: ReserveCacheRequest = {
+    const url = new URL('_apis/artifactcache/caches', this.#baseUrl);
+    const reqBody = JSON.stringify({
       key,
       version,
       cacheSize,
-    };
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const response: ITypedResponseWithError<ReserveCacheResponse> =
-      await retryTypedResponse('reserveCache', async () =>
-        httpClient.postJson(
-          /** For docs on this API, see {@link FakeGitHubActionsCacheServer} */
-          `${this.#baseUrl}_apis/artifactcache/caches`,
-          request
-        )
-      );
-    if (response.statusCode === /* Conflict */ 409) {
-      // This cache entry has already been reserved, so we can't write to it.
+    });
+    const {req, resPromise} = this.#request(url, {
+      method: 'POST',
+      headers: {
+        'content-type': 'application/json',
+      },
+    });
+    req.end(reqBody);
+    const res = await resPromise;
+
+    if (isOk(res)) {
+      const resData = JSON.parse(await readBody(res)) as {
+        cacheId: number;
+      };
+      return resData.cacheId;
+    }
+
+    if (res.statusCode === /* Conflict */ 409) {
       return undefined;
     }
-    if (response.statusCode === /* Too Many Requests */ 429) {
+
+    if (res.statusCode === /* Too Many Requests */ 429) {
       this.#onRateLimit(script);
       return undefined;
     }
-    if (
-      !isSuccessStatusCode(response.statusCode) ||
-      response.error !== undefined ||
-      response.result?.cacheId === undefined
-    ) {
-      throw new Error(
-        `Error reserving cache entry: ${response.statusCode} ${
-          response.error?.message ?? '<no error message>'
-        }`
-      );
-    }
-    return response.result.cacheId;
+
+    throw new Error(
+      `GitHub Cache reserve HTTP ${String(
+        res.statusCode
+      )} error: ${await readBody(res)}`
+    );
   }
 }
 
 class GitHubActionsCacheHit implements CacheHit {
+  #script: ScriptReference;
   #url: string;
-  #compressionMethod: CompressionMethod;
   #applied = false;
-  #emptyDirectoriesManifestPath: string;
 
-  constructor(
-    location: string,
-    compressionMethod: CompressionMethod,
-    emptyDirectoriesManifestPath: string
-  ) {
+  constructor(script: ScriptReference, location: string) {
+    this.#script = script;
     this.#url = location;
-    this.#compressionMethod = compressionMethod;
-    this.#emptyDirectoriesManifestPath = emptyDirectoriesManifestPath;
   }
 
   async apply(): Promise<void> {
@@ -504,40 +506,104 @@ class GitHubActionsCacheHit implements CacheHit {
       throw new Error('GitHubActionsCacheHit.apply was called more than once');
     }
     this.#applied = true;
-    const archivePath = pathlib.join(
-      await cacheUtils.createTempDirectory(),
-      cacheUtils.getCacheFileName(this.#compressionMethod)
-    );
+    const tempDir = await makeTempDir(this.#script);
+    const tarballPath = pathlib.join(tempDir, 'cache.tgz');
     try {
-      // TODO(aomarks) We should recover from rate limits and other HTTP errors
-      // here, but we currently seem to just get an exception about the tarball
-      // being invalid so we can't really tell what's going on.
-      await downloadCache(this.#url, archivePath);
-      await extractTar(archivePath, this.#compressionMethod);
-      await this.#createEmptyDirectories();
+      // TODO(aomarks) Recover from rate limits and other HTTP errors.
+      await this.#download(tarballPath);
+      await this.#extract(tarballPath);
     } finally {
-      await cacheUtils.unlinkFile(archivePath);
+      await fs.rm(tempDir, {recursive: true});
     }
   }
 
-  async #createEmptyDirectories() {
-    let manifest;
-    try {
-      manifest = await fs.readFile(this.#emptyDirectoriesManifestPath, {
-        encoding: 'utf8',
-      });
-    } catch (error) {
-      const {code} = error as {code: string};
-      if (code === 'ENOENT') {
-        // No empty dirs manifest means no empty dirs which is no problem.
-        return;
-      }
-      throw error;
+  async #download(tarballPath: string): Promise<void> {
+    const {req, resPromise} = request(this.#url);
+    req.end();
+    const res = await resPromise;
+    if (!isOk(res)) {
+      throw new Error(
+        `GitHub Cache download HTTP ${String(res.statusCode)} error`
+      );
     }
-    const emptyDirs = JSON.parse(manifest) as string[];
-    await Promise.all([
-      ...emptyDirs.map((dir) => fs.mkdir(dir, {recursive: true})),
-      fs.unlink(this.#emptyDirectoriesManifestPath),
-    ]);
+    await new Promise<void>((resolve, reject) => {
+      const writeTarballStream = createWriteStream(tarballPath);
+      writeTarballStream.on('error', (error) => reject(error));
+      res.on('error', (error) => reject(error));
+      res.pipe(writeTarballStream);
+      writeTarballStream.on('close', () => {
+        resolve();
+      });
+    });
   }
+
+  #extract(tarballPath: string): Promise<void> {
+    return new Promise<void>((resolve, reject) => {
+      execFile(
+        'tar',
+        ['--extract', '--file', tarballPath, '--gzip', '-P'],
+        (error: unknown) => {
+          if (error != null) {
+            reject(`tar error: ${String(error)}`);
+          } else {
+            resolve();
+          }
+        }
+      );
+    });
+  }
+}
+
+function request(
+  url: URL | string,
+  options?: http.RequestOptions
+): {
+  req: http.ClientRequest;
+  resPromise: Promise<http.IncomingMessage>;
+} {
+  const opts = {
+    ...options,
+    headers: {
+      // https://github.com/actions/toolkit/blob/500d0b42fee2552ae9eeb5933091fe2fbf14e72d/packages/cache/src/internal/cacheHttpClient.ts#L67
+      'user-agent': 'actions/cache',
+      ...options?.headers,
+    },
+  };
+  let req!: http.ClientRequest;
+  const resPromise = new Promise<http.IncomingMessage>((resolve, reject) => {
+    req = https.request(url, opts, (res) => {
+      resolve(res);
+    });
+    req.on('error', (error) => {
+      reject(error);
+    });
+  });
+  return {req, resPromise};
+}
+
+function isOk(res: http.IncomingMessage): boolean {
+  return (
+    res.statusCode !== undefined &&
+    res.statusCode >= 200 &&
+    res.statusCode < 300
+  );
+}
+
+function readBody(res: http.IncomingMessage): Promise<string> {
+  const chunks: Buffer[] = [];
+  res.on('data', (chunk: Buffer) => {
+    chunks.push(chunk);
+  });
+  return new Promise((resolve, reject) => {
+    res.on('error', (error: Error) => {
+      reject(error);
+    });
+    res.on('end', () => {
+      resolve(Buffer.concat(chunks).toString());
+    });
+  });
+}
+
+function makeTempDir(script: ScriptReference): Promise<string> {
+  return fs.mkdtemp(pathlib.join(getScriptDataDir(script), 'temp'));
 }